changed the np dtype to float32 instead of float64. this is expected by SB3 algos. change only affects pointRobot for now, but should be done for all other environments and their observation and action spaces

+++ planarenvs/planarCommon/velEnv.py
@@ -5,11 +5,11 @@ from gym import spaces
 class VelEnv(object):
     def setSpaces(self):
         self.observation_space = spaces.Dict({
-            'x': spaces.Box(low=-self._limUpPos, high=self._limUpPos, dtype=np.float64),
-            'xdot': spaces.Box(low=-self._limUpVel, high=self._limUpVel, dtype=np.float64)
+            'x': spaces.Box(low=-self._limUpPos, high=self._limUpPos, dtype=np.float32),
+            'xdot': spaces.Box(low=-self._limUpVel, high=self._limUpVel, dtype=np.float32)
         })
         self.action_space = spaces.Box(
-            low=-self._limUpVel, high=self._limUpVel, dtype=np.float64
+            low=-self._limUpVel, high=self._limUpVel, dtype=np.float32
         )
 
     def integrate(self):
